@@ -1,33 +1,27 @@
 package com.ruoyi.web.controller.tool;
 
+import com.ruoyi.common.core.controller.BaseController;
+import com.ruoyi.common.core.domain.R;
+import com.ruoyi.common.utils.StringUtils;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.Parameters;
+import io.swagger.v3.oas.annotations.enums.ParameterIn;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import org.springframework.web.bind.annotation.*;
+
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import com.ruoyi.common.core.controller.BaseController;
-import com.ruoyi.common.core.domain.R;
-import com.ruoyi.common.utils.StringUtils;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiImplicitParam;
-import io.swagger.annotations.ApiImplicitParams;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-import io.swagger.annotations.ApiOperation;
 
 /**
  * swagger 用户测试方法
- * 
+ *
  * @author ruoyi
  */
-@Api("用户信息管理")
+@Tag(name = "用户信息管理")
 @RestController
 @RequestMapping("/test/user")
 public class TestController extends BaseController
@@ -38,7 +32,7 @@ public class TestController extends BaseController
         users.put(2, new UserEntity(2, "ry", "admin123", "15666666666"));
     }
 
-    @ApiOperation("获取用户列表")
+    @Operation(description="获取用户列表")
     @GetMapping("/list")
     public R<List<UserEntity>> userList()
     {
@@ -46,8 +40,8 @@ public class TestController extends BaseController
         return R.ok(userList);
     }
 
-    @ApiOperation("获取用户详细")
-    @ApiImplicitParam(name = "userId", value = "用户ID", required = true, dataType = "int", paramType = "path", dataTypeClass = Integer.class)
+    @Operation(description="获取用户详细")
+    @Parameters({@Parameter(name = "userId", description = "用户ID", required = true,in = ParameterIn.PATH,schema =@Schema(implementation=Integer.class))})
     @GetMapping("/{userId}")
     public R<UserEntity> getUser(@PathVariable Integer userId)
     {
@@ -61,12 +55,12 @@ public class TestController extends BaseController
         }
     }
 
-    @ApiOperation("新增用户")
-    @ApiImplicitParams({
-        @ApiImplicitParam(name = "userId", value = "用户id", dataType = "Integer", dataTypeClass = Integer.class),
-        @ApiImplicitParam(name = "username", value = "用户名称", dataType = "String", dataTypeClass = String.class),
-        @ApiImplicitParam(name = "password", value = "用户密码", dataType = "String", dataTypeClass = String.class),
-        @ApiImplicitParam(name = "mobile", value = "用户手机", dataType = "String", dataTypeClass = String.class)
+    @Operation(description="新增用户")
+    @Parameters({
+            @Parameter(name = "userId", description = "用户id", schema =@Schema(implementation=Integer.class)),
+            @Parameter(name = "username", description = "用户名称", schema =@Schema(implementation=String.class)),
+            @Parameter(name = "password", description = "用户密码",  schema =@Schema(implementation=String.class)),
+            @Parameter(name = "mobile", description = "用户手机", schema =@Schema(implementation=String.class))
     })
     @PostMapping("/save")
     public R<String> save(UserEntity user)
@@ -79,7 +73,7 @@ public class TestController extends BaseController
         return R.ok();
     }
 
-    @ApiOperation("更新用户")
+    @Operation(description="更新用户")
     @PutMapping("/update")
     public R<String> update(@RequestBody UserEntity user)
     {
@@ -96,8 +90,8 @@ public class TestController extends BaseController
         return R.ok();
     }
 
-    @ApiOperation("删除用户信息")
-    @ApiImplicitParam(name = "userId", value = "用户ID", required = true, dataType = "int", paramType = "path", dataTypeClass = Integer.class)
+    @Operation(description="删除用户信息")
+    @Parameter(name = "userId", description = "用户ID", required = true,  in = ParameterIn.PATH, schema =@Schema(implementation=Integer.class))
     @DeleteMapping("/{userId}")
     public R<String> delete(@PathVariable Integer userId)
     {
@@ -113,19 +107,19 @@ public class TestController extends BaseController
     }
 }
 
-@ApiModel(value = "UserEntity", description = "用户实体")
+@Schema(title = "UserEntity", description = "用户实体")
 class UserEntity
 {
-    @ApiModelProperty("用户ID")
+    @Schema(description = "用户ID")
     private Integer userId;
 
-    @ApiModelProperty("用户名称")
+    @Schema(description ="用户名称")
     private String username;
 
-    @ApiModelProperty("用户密码")
+    @Schema(description ="用户密码")
     private String password;
 
-    @ApiModelProperty("用户手机")
+    @Schema(description ="用户手机")
     private String mobile;
 
     public UserEntity()
